+++ packages/plugins/thorchain/CHANGELOG.md
@@ -1,5 +1,67 @@
 # @swapkit/plugin-thorchain
 
+## 1.1.28
+
+### Patch Changes
+
+- [#981](https://github.com/thorswap/SwapKit/pull/981) [`71e0e5a`](https://github.com/thorswap/SwapKit/commit/71e0e5acf6d771bee816c9da3744d56e385392e6) Thanks [@towanTG](https://github.com/towanTG)! - Adds recipient address validation for tc plugin
+
+## 1.1.27
+
+### Patch Changes
+
+- Updated dependencies [[`5d44a3c`](https://github.com/thorswap/SwapKit/commit/5d44a3c48dd458ce4f1265934a9f87eabbc7886a)]:
+  - @swapkit/toolbox-cosmos@1.3.0
+  - @swapkit/helpers@1.8.0
+  - @swapkit/api@1.6.6
+  - @swapkit/toolbox-evm@1.3.6
+  - @swapkit/toolbox-substrate@1.2.6
+  - @swapkit/toolbox-utxo@1.1.11
+
+## 1.1.26
+
+### Patch Changes
+
+- Updated dependencies [[`17f28d9`](https://github.com/thorswap/SwapKit/commit/17f28d901f35116960766f8c872a91baac67bd6a)]:
+  - @swapkit/toolbox-cosmos@1.2.1
+  - @swapkit/helpers@1.7.1
+  - @swapkit/api@1.6.5
+  - @swapkit/toolbox-evm@1.3.5
+  - @swapkit/toolbox-substrate@1.2.5
+  - @swapkit/toolbox-utxo@1.1.10
+
+## 1.1.25
+
+### Patch Changes
+
+- Updated dependencies [[`97b8d3f`](https://github.com/thorswap/SwapKit/commit/97b8d3fe31987cd0813c847159b8127087dada10)]:
+  - @swapkit/toolbox-cosmos@1.2.0
+  - @swapkit/helpers@1.7.0
+  - @swapkit/api@1.6.4
+  - @swapkit/toolbox-evm@1.3.4
+  - @swapkit/toolbox-substrate@1.2.4
+  - @swapkit/toolbox-utxo@1.1.9
+
+## 1.1.24
+
+### Patch Changes
+
+- Updated dependencies [[`59447e0`](https://github.com/thorswap/SwapKit/commit/59447e00ea8f32e30fbd61b9aa6e07314b42d6b6)]:
+  - @swapkit/helpers@1.6.3
+  - @swapkit/api@1.6.3
+  - @swapkit/toolbox-cosmos@1.1.5
+  - @swapkit/toolbox-evm@1.3.3
+  - @swapkit/toolbox-substrate@1.2.3
+  - @swapkit/toolbox-utxo@1.1.8
+
+## 1.1.23
+
+### Patch Changes
+
+- Updated dependencies [[`cf2ccbb`](https://github.com/thorswap/SwapKit/commit/cf2ccbb016d77ea36bc0a2831daba0b9aaa96685)]:
+  - @swapkit/toolbox-cosmos@1.1.4
+  - @swapkit/helpers@1.6.2
+
 ## 1.1.22
 
 ### Patch Changes